backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "The API for Seecho AI"
authors = ["PancakePuncher"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.1"
pyseto = "^1.7.3"
uvicorn = "^0.23.2"
python-multipart = "^0.0.6"
argon2-cffi = "^21.3.0"
tortoise-orm = {extras = ["asyncmy"], version = "^0.19.3"}
pydantic-settings = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
